requirements.txt
Flask>=2.2
requests>=2.28
Flask-Caching>=1.11
pytest>=7.0
python-dotenv>=1.0
